test: Setup connection globally
