+++ server/src/test/initState.ts
@@ -1,0 +1,22 @@
+export default {
+  user: {
+    id: 1,
+    username: "test1",
+    email: "test@test.aom",
+    thumbnail: "thumbnail",
+    socialAccount: {
+      provider: "google",
+      socialId: "123",
+    },
+  },
+  zettel: {
+    id: 3,
+    version: 1,
+    uuid: "5724ae30-1383-45d6-bf4c-6f46e8768079",
+    content: "testcontent",
+    contentType: "md",
+    title: "testtitle",
+    // tags: ["t1", "t2"],
+    userId: 1,
+  },
+};
