+++ server/src/test/jest-global-teardown.ts
@@ -1,0 +1,6 @@
+import { getConnection } from "typeorm";
+
+module.exports = async () => {
+  const conn = getConnection();
+  await conn.close();
+};
